Add support for the DraCor Schema
This is a major revision renaming the action inputs and adding the 'schema'
input allowing to choose between the TEI_all and DraCor schemas.
With this revision we also support other schema versions than the bundled
ones. Recent versions are still bundled but if an unbundled one is
requested, the action will attempt to download it from the known location.